refactor: extract graphql editor pattern
