+++ src/components/patterns/index.js
@@ -15,6 +15,7 @@ import Aside from './Aside/Aside'
 import Header from './Header/Header'
 import Caption from './Caption/Caption'
 import Announcement from './Announcement/Announcement'
+import GraphqlEditor from './GraphqlEditor/GraphqlEditor'
 
 export {
   Announcement,
@@ -25,6 +26,7 @@ export {
   CookiesPolicy,
   DemoLinks,
   Footer,
+  GraphqlEditor,
   Grid,
   Header,
   Layout,
